Create 3n + 1
Consider the following algorithm to generate a sequence of numbers. Start with an integer n. If n is even, divide it by 2. If n is odd, multiply it by 3 and add 1.

Repeat this process with the new value of n, ending when n = 1.

For example, the following sequence of numbers will be generated when n is 22:

22 11 34 17 52 26 13 40 20 10 5 16 8 4 2 1

Although there is still no proof, mathematicians believe that this algorithm always ends with n = 1, for any integer n. Well, for this problem here at Huxley, this property holds for any integer less than 1,000,000.

For an input n, the "cycle size" of n is the number of numbers generated up to 1 (including 1 itself). In the example above, the "cycle size" of 22 is 16.

Given two numbers i and j, determine the maximum "cycle size" of all numbers between i and j, including both i and j.

Formato d