Create Family Photo
Ambrosina is a very peculiar photographer. She only accepts taking photos of people if they are in groups of exactly 4. All that because Ambrosina has a weird ordination habit. For her, the shortest person must always stay at the left side, the second shortest at the right side, the second tallest stays in the middle, after the shortest, and the tallest must be after her (to the left of the second shortest).