Create Elevator
Going Very Comfortably (GVC) is a company with more than 50 years of tradition and experience manufacturing elevators. Every GVC project follows the strictest safety standards; however, a series of accidents with their elevators ruined the company's reputation. After studying those tragedies, the engineers concluded that, in several occurrences, the accident was caused by the presence of too many passengers in the elevator.

So, GVC decided to supervise more strictly the use of its elevators: sensors which are able to detect how many people come in and come out in each floor were installed in each elevator door. GVC keeps the sensor's logs for a whole day of operation. The logs always start in empty state.

They know that people are polite and always let all the passengers that will leave on one floor exit the elevator before other passengers enter the elevator. However, it is still difficult to determine if the maximum elevator capacity has been exceeded or not.

Write a program that, given a sequence of sensor readings and the maximum elevator capacity, determine whether the maximum elevator capacity was exceeded at any moment.