Create Highest Number Challenge
Leonardo is a very creative kid. He loves to create challenges for his schoolmates.

His last challenge is as follows: several numbers are said out loud, when the number 0 (zero) is said then the challenge ends and his colleagues must immediately tell you which was the highest number. Leonardo has a hard time checking whether the answer given by his colleagues is correct or not, since the sequence of numbers is often long. For this reason, he decided to ask for your help.

Your task is to write a program that, given a sequence of positive integers ending with 0 (zero), prints the highest number in the sequence.