Create Stop Pollution
Due to the high pollution ratios, a city decided to fine those houses where there are more than two vehicles. The fine will cost R$ 12.89 per month for each extra vehicle. Write a program to receive the amount of vehicles of many residences as input, until the value of 999 comes in, and then shows the collected monthly amount with the fines and the number of fined houses.