Create  Black Friday Discounts
Black Friday is a popular sale that takes place in many stores after American Thanksgiving. Genésio does not believe much in the advertisements and thinks that it is all hoaxing and that the discounts are not so good.

So he decided to do a search in an online store. He registered the price of 5 products a week before and on the day of the sale to see if the announced 20% discount had actually been applied.

Write a program that receives the original and discounted prices of each product as input, and displays how many actually were rebated at least 20% at the end.