Create Christmas Tree
This year, Robson wants to build a nice Christmas Tree in the company which he works. He agreed with the other 20 employees of his department to buy everything and share the amount afterwards. Searching through a store, he found a perfect tree and three kinds of ornaments with good prices. 

Write a program which receives the cost of the tree, how many ornaments were bought and the corresponding cost for each type as input, and calculates the total amount spent and how much will be paid by each employee