Create Games' Price
Ambrósio loves to play video games and like a professional player, he decided to create a relation between the price of games and their difficulty.

He found out that every year the games become cheaper, but there are some that, due to their difficulty, keep a somewhat higher price, because they are more challenging and bought more often.

Ambrósio separated the games' difficulties from 0 to 5, and calculated them according to the current game prices, finding out rules that measure with good precision how much the same game will cost next year. The rules are:

If the game's difficulty is 0, the price will drop 25%

If the game's difficulty is 1, the price will drop 20%

If the game's difficulty is 2, the price will drop 18%

If the game's difficulty is 3, the price will drop 15%

If the game's difficulty is 4, the price will drop 12%

If the game's difficulty is 5, the price will drop 10%



However, there is a problem. When the game is very devalued, that is, its price is already below or equal to R$100.00 in the current year, the price will only drop half of what it should, until R$45 minimum.

Knowing that you are a programmer, Ambrósio asked for your help to create a program to calculate the price of N games after 3 years which is the maximum time he is willing to wait.